6_1-Ft_transcendence/backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.0
channels-redis==4.2.1
pillow==11.1.0
daphne==4.1.2
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
pyjwt==2.10.1
httplib2==0.18.1
hyperlink==21.0.0
psycopg2-binary==2.9.10
pyOpenSSL==24.3.0
redis==5.2.0
requests